requirements.txt
flask
psycopg2-binary
fpdf
gunicorn